Move HotelCell test helpers to a separated file
